requirements.txt
python-dotenv
droidrun[google]
llama-index-llms-google-genai
colorama